ported dexnav from e-expansion

+++ include/bg.h
@@ -91,7 +91,7 @@ u16 GetBgAttribute(u8 bg, u8 attributeId);
 u32 ChangeBgX(u8 bg, u32 value, u8 op);
 u32 GetBgX(u8 bg);
 u32 ChangeBgY(u8 bg, u32 value, u8 op);
-u32 ChangeBgY_ScreenOff(u8 bg, u32 value, u8 op);
+s32 ChangeBgY_ScreenOff(u32 bg, s32 value, u8 op);
 u32 GetBgY(u8 bg);
 void SetBgAffine(u8 bg, u32 srcCenterX, u32 srcCenterY, s16 dispCenterX, s16 dispCenterY, s16 scaleX, s16 scaleY, u16 rotationAngle);
 u8 AdjustBgMosaic(u8 value, u8 mode);
